src/requirements.txt
graphviz
matplotlib
opencv-python
pandas
pillow
pydot
pyyaml
statsmodels
seaborn
sklearn
tensorflowjs
tensorflow
tensorflow_datasets
# tensorflow-gpu

imageio==2.5.0
pytest==5.0.0
seaborn==0.9.0
pycocotools==2.0.0
labelImg==1.8.3
lxml==4.3.4
